+++ packages/etlantic-pyspark/src/etlantic_pyspark/provider.py
@@ -14,7 +14,7 @@ from etlantic.spark.provider import (
     SparkSessionRequest,
 )
 
-__version__ = "0.15.0"
+__version__ = "0.16.0"
 
 
 class LocalSparkProvider: